telegram bugfix

+++ content/saloos_tg/sarshomar_bot/controller.php
@@ -59,7 +59,7 @@ class controller extends \lib\mvc\controller
 		 * start hooks and run telegram session from db
 		 */
 		bot::hook();
-		\lib\main::$controller->model()->user_id = bot::$user_id;
+		\lib\main::$controller->model()->user_id = (int) bot::$user_id;
 		$language = \lib\db\users::get_language((int) bot::$user_id);
 		if(empty($language) || !$language)
 		{
